+++ bedcode-desktop/src/router/index.ts
@@ -29,28 +29,29 @@ const router = createRouter({
       name: 'settings',
       component: () => import('@/views/SettingsView.vue'),
     },
-    {
-      path: '/plugins',
-      name: 'plugins',
-      component: () => import('@/views/PluginsView.vue'),
-    },
-    {
-      path: '/plugins/:id/config',
-      name: 'plugin-config',
-      component: () => import('@/views/PluginConfigView.vue'),
-    },
-    {
-      path: '/plugin/sidebar/:pluginId/:viewId',
-      name: 'plugin-sidebar-view',
-      component: () => import('@/plugin/components/PluginViewHost.vue'),
-      props: true,
-    },
-    {
-      path: '/plugin/toolbox/:pluginId/:viewId',
-      name: 'plugin-toolbox-view',
-      component: () => import('@/plugin/components/PluginViewHost.vue'),
-      props: true,
-    },
+    // TODO: 插件功能暂未上线
+    // {
+    //   path: '/plugins',
+    //   name: 'plugins',
+    //   component: () => import('@/views/PluginsView.vue'),
+    // },
+    // {
+    //   path: '/plugins/:id/config',
+    //   name: 'plugin-config',
+    //   component: () => import('@/views/PluginConfigView.vue'),
+    // },
+    // {
+    //   path: '/plugin/sidebar/:pluginId/:viewId',
+    //   name: 'plugin-sidebar-view',
+    //   component: () => import('@/plugin/components/PluginViewHost.vue'),
+    //   props: true,
+    // },
+    // {
+    //   path: '/plugin/toolbox/:pluginId/:viewId',
+    //   name: 'plugin-toolbox-view',
+    //   component: () => import('@/plugin/components/PluginViewHost.vue'),
+    //   props: true,
+    // },
     {
       path: '/terminal-window/:id',
       name: 'terminal-window',