+++ bedcode-mobile/src/components/FileViewerModal.vue
@@ -78,6 +78,7 @@
   <!-- Settings Modal -->
   <CodeViewerSettingsModal
     :visible="showSettings"
+    :z-index="60"
     @close="showSettings = false"
     @confirm="showSettings = false"
   />